1. Ответственности разбиты по файлам и организованы 2. в MovieQuizViewController.swift были внесены корректировки для его совместной работы с другими файлами проекта.
