+++ MovieQuiz/Models/QuizQuestion.swift
@@ -1,0 +1,18 @@
+//
+//  QuizQuestion.swift
+//  MovieQuiz
+//
+//  Created by Алишер Дадаметов on 10.04.2023.
+//
+
+import Foundation
+
+
+struct QuizQuestion {
+    let image: String
+    let text: String
+    let correctAnswer: Bool
+        }
+    
+
+
