+++ MovieQuiz/Models/QuizResultsViewModel.swift
@@ -1,0 +1,16 @@
+//
+//  QuizResultsViewModel.swift
+//  MovieQuiz
+//
+//  Created by Алишер Дадаметов on 10.04.2023.
+//
+
+import Foundation
+
+struct QuizResultsViewModel {
+    let title: String
+    let text: String
+    let buttonText: String
+    }
+    
+
