+++ MovieQuiz/Models/QuizStepViewModel.swift
@@ -1,0 +1,19 @@
+//
+//  QuizStepViewModel.swift
+//  MovieQuiz
+//
+//  Created by Алишер Дадаметов on 10.04.2023.
+//
+import Foundation
+import UIKit
+
+
+struct QuizStepViewModel{
+    let image: UIImage
+    let question: String
+    let questionNumber: String
+    }
+
+
+
+
